backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
psycopg2-binary==2.9.10
djangorestframework-simplejwt==5.5.1
python-dotenv==1.1.1